Refactor GC compaction: Move module to lib/
Move GcCompactThread module from config/initializers to lib/ for better
organization and separation of concerns.

Changes:

- Created lib/gc_compact_thread.rb with the GcCompactThread module
- Added start_background_thread method to encapsulate initialization
- Simplified config/initializers/gc_compact_thread.rb (65 lines -> 12 lines)
- Rails autoloads from lib/ so we don't need a require_relative

The initializer now contains only initialization logic, while all business
logic resides in the testable module file following Rails conventions.
